safari-534.53-branch/LayoutTests/fast/workers/storage/change-version-sync-expected.txt

Test that version updates in sync databases work correctly.
PASS: db.version is 3 as expected.
